+++ test/3D/config/dvec.hpp
@@ -1,15 +1,15 @@
 #pragma once
 
-struct dvec3{
+struct vec3{
   double x,y,z;
-  bool operator==(const dvec3& v) const{
+  bool operator==(const vec3& v) const{
     return (v.x==x && v.y==y && v.z==z);
   }
 };
 
-struct dvec4{
+struct vec4{
   double x,y,z,w;
-  bool operator==(const dvec4& v) const{
+  bool operator==(const vec4& v) const{
     return (v.x==x && v.y==y && v.z==z && v.w == w);
   }
 };
